merge

+++ android-app-helloworld-master/app/src/main/java/com/navisens/demo/android_app_helloworld/database_obj/Path.java
@@ -11,7 +11,7 @@ import androidx.room.PrimaryKey;
 public class Path {
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "path_id")
-    @NonNull public int pid;
+    @NonNull public long pid;
 
     @ColumnInfo(name = "name")
     @NonNull public String name;
